Remove separation between x86_64 and i386 for node and appliance kickstarts
In Fedora 8 we needed to maintain two kickstarts (one for each arch) for
the managed node and the wui appliance since Fedora 8 would install the
i386 versions of RPMs in the x86_64 OS.  Fedora 9 is smarter about this
and only installs the package specific to the arch.  Since the only
difference between the two kickstarts was the removal of specific
i386/i686 packages, this can be simplified.

Signed-off-by: Perry Myers <pmyers@redhat.com>