Add .noarch to ovirt-wui package in wui-appliance kickstart
ovirt-wui package moved from x86_64/i386 to noarch since there is no
longer any C code in there (all scripts now).  When this was done
we ran into a problem with older versions of ovirt-wui with
x86_64 taking precedence over newer versions of ovirt-wui.noarch.
This doesn't happen on i386, so it is probably due to multilib in
yum.

Signed-off-by: Perry Myers <pmyers@redhat.com>