includes basic search functionality.
For this to work, you've got to have xapian-core and xapian-bindings-ruby installed. The latter is currently only available in updates-testing, but apevec is adding it to the ovirt repo for now. In addition, there's a new migration required here, so you'll nede to run rake db:migrate.

Currently we do a full reindex when restarting mongrel, and incremental index updates are not yet in place. Also missing is permission checking.
I've included basic fields in the search parameters -- we should look through and determine if any additional attributes/associations need to be included.

Also, the styling is _mostly_ consistent with tallen's mockups. Notable differences here:
1) inline search field needs css and possibly a new image.
2) I need "selected" and "unselected" images for the type pulldown -- currently I'm just using a text "X" for selected and &nbsp; for unselected
3) No "smart" pool column or pulldown, since that functionality is not yet implemented
4) I've added a ranking column to the results
5) details pane works here just like it does for hosts, vms, etc. on other pages.