Fix broken js cause by svg widget. (revised)
This just disables the calls to svg widget within any flexigrids, which is the cause of checkboxes not appearing, details not showing, and other such unpleasantness.  Next step would be to replace these bits with css bar graphs, which will hopefully make it in shortly, but as this breakage is annoying, thought I would do a quick fix.  Left it as commented out since the very next step is to replace with css and I want to easily see what I am replacing.